Some reorganization and marios suggestions
Made it possible to stack trained or not trained input and output mappings on to each other
